% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpomological-package.R
\docType{package}
\name{ggpomological-package}
\alias{ggpomological}
\alias{ggpomological-package}
\title{A Pomological ggplot2 Theme}
\description{
This package provides a ggplot2 theme inspired by the
\href{https://usdawatercolors.nal.usda.gov/pom}{USDA Pomological Watercolors collection}
and by Aron Atkins's (\href{https://twitter.com/aronatkins]}{\@aronatkins})
\href{https://github.com/rstudio/rstudio-conf/tree/master/2018/Fruit_For_Thought--Aron_Atkins}{talk on parameterized RMarkdown}
at \href{https://www.rstudio.com/conference/}{rstudio::conf 2018}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/ggpomological}
  \item Report bugs at \url{https://github.com/gadenbuie/ggpomological/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{g.adenbuie@gmail.com}

}
