% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_pomological.R
\name{paint_pomological}
\alias{paint_pomological}
\title{Paint a ggpomological watercolor}
\usage{
paint_pomological(pomo_gg, width = 800, height = 500, pointsize = 16,
  outfile = NULL, pomological_background = pomological_images("background"),
  pomological_overlay = pomological_images("overlay"), ...)
}
\arguments{
\item{pomo_gg}{A pomologically styled ggplot2 object. See \code{\link[=theme_pomological]{theme_pomological()}}}

\item{width}{Width of output image in pixels}

\item{height}{Height of output image in pixels}

\item{pointsize}{Text size for plot text}

\item{outfile}{Optional name for output file if you'd like to save the image}

\item{pomological_background}{Paper image, defaults to paper texture provided
by ggpomological.}

\item{pomological_overlay}{Overlay texture. Set to \code{NULL} for no texture.}

\item{...}{Arguments passed on to \code{magick::image_graph}
\describe{
  \item{res}{resolution in pixels}
  \item{clip}{enable clipping in the device. Because clipping can slow things down
a lot, you can disable it if you don't need it.}
  \item{antialias}{TRUE/FALSE: enables anti-aliasing for text and strokes}
}}
}
\description{
Uses \link{magick} to paint a pomological watercolor. (Paints your plot onto a
pomological watercolor style paper, with texture overlay.)
}
\references{
https://usdawatercolors.nal.usda.gov/pom
}
\seealso{
\link{theme_pomological}
}
