% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pomological.R
\name{theme_pomological}
\alias{theme_pomological}
\alias{theme_pomological_nobg}
\alias{theme_pomological_plain}
\title{Pomological Theme}
\usage{
theme_pomological(base_family = "Homemade Apple", base_size = 16,
  text.color = pomological_base$dark_blue,
  plot.background.color = pomological_base$paper,
  panel.border.color = pomological_base$light_line, with.panel.grid = FALSE,
  panel.grid.color = pomological_base$light_line,
  panel.grid.linetype = "dashed",
  axis.text.color = pomological_base$medium_line, axis.text.size = base_size
  * 14/16, base_theme = ggplot2::theme_minimal())

theme_pomological_nobg(...)

theme_pomological_plain(...)
}
\arguments{
\item{base_family, base_size}{Base text family and size}

\item{text.color}{Color of all text (except axis text, see \code{axis.text.color})}

\item{plot.background.color}{Color of plot background, passed to \code{plot.background}}

\item{panel.border.color}{Color of plot panel border}

\item{with.panel.grid}{If \code{FALSE} gridlines in plot are removed}

\item{panel.grid.color, panel.grid.linetype}{Color and linetype of panel grid, passed to \code{panel.grid}}

\item{axis.text.color, axis.text.size}{Color and size of axis text}

\item{base_theme}{Starting theme of plot, default is
\code{\link[ggplot2:theme_minimal]{ggplot2::theme_minimal()}}. Any elements set by \code{theme_pomological()} will
overwrite the \code{base_theme} unless the specific parameter is explicitly set
to \code{NULL}.}
}
\description{
\link{ggplot2} plot theme based on the USDA Pomological Watercolors paintings.
}
\section{Functions}{
\itemize{
\item \code{theme_pomological_nobg}: Pomological theme with white (transparent) background

\item \code{theme_pomological_plain}: A "plain" pomological theme with white
background and normal fonts.
}}

\section{Fonts}{

Complete the pomological watercolor theme with a handwriting or cursive font.
The following fonts from \href{https://fonts.google.com}{Google Fonts} work well.
Visit the links below to install on your system.
\itemize{
\item \href{https://fonts.google.com/specimen/Homemade+Apple/}{Homemade Apple}
\item \href{https://fonts.google.com/specimen/Amatic+SC/}{Amatic SC}
\item \href{https://fonts.google.com/specimen/Mr+Bedfort/}{Mr. Bedfort}
}

Fonts with R are notoriously tricky, so these may not work well for you. If
you have installed the fonts but they aren't showing up or working, you can
always try running \code{extrafont::font_import()} or \code{extrafont::load_fonts()} in
the session or RMarkdown document. Or you can use \code{\link[=theme_pomological_plain]{theme_pomological_plain()}}.
}

\examples{
library(ggplot2)
basic_iris_plot <- ggplot(iris) +
  aes(x = Sepal.Length, y = Sepal.Width, color = Species) +
  geom_point(size = 2) +
  # with pomological color scale
  scale_color_pomological()

# Pomological Theme  
basic_iris_plot + 
  theme_pomological() 

# Don't change panel grid color
basic_iris_plot + 
  theme_pomological(
    panel.grid.color = NULL
  )
  
# White background
basic_iris_plot +
  theme_pomological_nobg()
  
# Plain plot without font or background
basic_iris_plot +
  theme_pomological_plain()

}
\references{
https://usdawatercolors.nal.usda.gov/pom
}
\seealso{
\link[ggplot2:theme]{ggplot2::theme}
}
