% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pomological.R
\name{theme_pomological}
\alias{theme_pomological}
\alias{theme_pomological_nobg}
\alias{theme_pomological_plain}
\alias{theme_pomological_fancy}
\title{Pomological Theme}
\usage{
theme_pomological(
  base_family = NULL,
  base_size = 11,
  text.color = pomological_base$dark_blue,
  plot.background.color = pomological_base$paper,
  panel.border.color = pomological_base$light_line,
  with.panel.grid = FALSE,
  panel.grid.color = pomological_base$light_line,
  panel.grid.linetype = "dashed",
  axis.text.color = pomological_base$medium_line,
  axis.text.size = base_size * 3/4,
  base_theme = ggplot2::theme_minimal()
)

theme_pomological_nobg(..., plot.background.color = "transparent")

theme_pomological_plain(
  base_family = "",
  base_size = 11,
  plot.background.color = "transparent",
  ...
)

theme_pomological_fancy(base_family = "Homemade Apple", base_size = 16, ...)
}
\arguments{
\item{base_family}{Base text family. See \strong{Fonts} in \code{\link[=theme_pomological]{theme_pomological()}}
for some examples from Google Fonts options, including \code{"Mr De Haviland"},
\code{"Homemade Apple"}, \code{"Marck Script"}, and \code{"Mr. Bedfort"}. For the
authentic pomological look, use \code{"Homemade Apple"} or \code{"Mr De Haviland"}.
Set to \code{NULL} or use \code{\link[=theme_pomological_plain]{theme_pomological_plain()}} for no change to fonts.}

\item{base_size}{Base text size}

\item{text.color}{Color of all text (except axis text, see \code{axis.text.color})}

\item{plot.background.color}{Color of plot background, passed to \code{plot.background}}

\item{panel.border.color}{Color of plot panel border}

\item{with.panel.grid}{If \code{FALSE} gridlines in plot are removed}

\item{panel.grid.color, panel.grid.linetype}{Color and linetype of panel grid, passed to \code{panel.grid}}

\item{axis.text.color, axis.text.size}{Color and size of axis text}

\item{base_theme}{Starting theme of plot, default is
\code{\link[ggplot2:theme_minimal]{ggplot2::theme_minimal()}}. Any elements set by \code{theme_pomological()} will
overwrite the \code{base_theme} unless the specific parameter is explicitly set
to \code{NULL}.}
}
\description{
\link{ggplot2} plot theme based on the USDA Pomological Watercolors paintings.
}
\section{Functions}{
\itemize{
\item \code{theme_pomological_nobg}: Pomological theme with white (transparent) background

\item \code{theme_pomological_plain}: A "plain" pomological theme with white
background and normal fonts.

\item \code{theme_pomological_fancy}: A "fancy" pomological theme with fancy fonts
}}

\section{Fonts}{

Complete the pomological watercolor theme with a handwriting or cursive font.
The following fonts from \href{https://fonts.google.com}{Google Fonts} work well.
Visit the links below to install on your system.
\itemize{
\item \href{https://fonts.google.com/specimen/Homemade+Apple/}{Homemade Apple}
\item \href{https://fonts.google.com/specimen/Mr+De+Haviland}{Mr. De Haviland}
\item \href{https://fonts.google.com/specimen/Marck+Script/}{Marck Script}
\item \href{https://fonts.google.com/specimen/Mr+Bedfort/}{Mr. Bedfort}
}

Fonts with R are notoriously tricky, so these may not work well for you. If
you have installed the fonts but they aren't showing up or working, you can
always try running \code{extrafont::font_import()} or \code{extrafont::load_fonts()} in
the session or RMarkdown document. Or you can use \code{\link[=theme_pomological_plain]{theme_pomological_plain()}}.
}

\examples{
library(ggplot2)

# Prep msleep data
msleep <- ggplot2::msleep[, c("vore", "sleep_rem", "sleep_total")]
msleep <- msleep[complete.cases(msleep), ]
msleep$vore <- paste0(msleep$vore, "vore")

# Base plot
basic_msleep_plot <- ggplot(msleep) +
  aes(x = sleep_rem, y = sleep_total, color = vore) +
  geom_point(size = 2) +
  labs(color = NULL)

# Just your standard ggplot
basic_msleep_plot 

# With pomological colors
basic_msleep_plot <- basic_msleep_plot + scale_color_pomological()
basic_msleep_plot

# With pomological theme
basic_msleep_plot + theme_pomological()

# With transparent background
basic_msleep_plot + theme_pomological_plain() 

# Or with "fancy" pomological settings
pomological_msleep <- basic_msleep_plot + theme_pomological_fancy()

# Painted!
paint_pomological(pomological_msleep, res = 110)

}
\references{
https://usdawatercolors.nal.usda.gov/pom
}
\seealso{
\link[ggplot2:theme]{ggplot2::theme}
}
