% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pomological.R
\name{theme_pomological}
\alias{theme_pomological}
\alias{theme_pomological_nobg}
\title{Pomological Theme}
\usage{
theme_pomological(base_family = "Homemade Apple", base_size = 16,
  text.color = NULL, plot.background.color = NULL,
  panel.grid.color = NULL, panel.grid.linetype = "dashed",
  axis.text.color = NULL, axis.text.size = base_size * 14/16,
  base_theme = ggplot2::theme_minimal())

theme_pomological_nobg(...)
}
\arguments{
\item{base_family}{Base text family}

\item{base_size}{Base text size}

\item{text.color}{Color of all text (except axis text, see \code{axis.text.color})}

\item{plot.background.color}{Color of plot background, passed to \code{plot.background}}

\item{panel.grid.color}{Color of panel grid, passed to \code{panel.grid}}

\item{panel.grid.linetype}{Linetype of panel grid, passed to \code{panel.grid}}

\item{axis.text.color}{Color of axis text}

\item{axis.text.size}{Size of axis text}

\item{base_theme}{Starting theme of plot, default is
\code{\link[ggplot2:theme_minimal]{ggplot2::theme_minimal()}}. Any elements set by \code{theme_pomological()} will
overwrite the \code{base_theme} unless the specific parameter is explicitly set
to \code{NULL}.}
}
\description{
\link{ggplot2} plot theme based on the USDA Pomological Watercolors paintings.
}
\section{Functions}{
\itemize{
\item \code{theme_pomological_nobg}: Pomological theme with white (transparent) background
}}

\section{Fonts}{

Complete the pomological watercolor theme with a handwriting or cursive font.
The following fonts from \href{https://fonts.google.com}{Google Fonts} work well.
Visit the links below to install on your system.
\itemize{
\item \href{https://fonts.google.com/specimen/Homemade+Apple/}{Homemade Apple}
\item \href{https://fonts.google.com/specimen/Amatic+SC/}{Amatic SC}
\item \href{https://fonts.google.com/specimen/Mr+Bedfort/}{Mr. Bedfort}
}
}

\examples{
library(ggplot2)
basic_iris_plot <- ggplot(iris) +
  aes(x = Sepal.Length, y = Sepal.Width, color = Species) +
  geom_point(size = 2)

# Pomological Theme  
basic_iris_plot + theme_pomological()

# Don't change panel grid color
basic_iris_plot + 
  theme_pomological(
    panel.grid.color = NULL
  )
  
# White background
basic_iris_plot +
  theme_pomological_nobg()

}
\references{
https://usdawatercolors.nal.usda.gov/pom
}
\seealso{
\link[ggplot2:theme]{ggplot2::theme}
}
